+++ vampytest/core/event_handling/__init__.py
@@ -1,3 +1,5 @@
+from .rendering_helpers import *
+
 from .base import *
 from .default import *
 from .default_output_writer import *
@@ -5,6 +7,8 @@ from .text_styling import *
 
 
 __all__ = (
+    *rendering_helpers.__all__,
+    
     *base.__all__,
     *default.__all__,
     *default_output_writer.__all__,