+++ vampytest/core/event_handling/default_output_writer.py
@@ -41,7 +41,7 @@ class OutputWriter(RichAttributeErrorBaseType):
         return self
     
     
-    def write_break_line(self, character="="):
+    def write_break_line(self, character = "="):
         """
         Writes a break line.