+++ vampytest/core/result/__init__.py
@@ -1,11 +1,10 @@
 from .failures import *
 
 from .result import *
-from .result_group import *
+
 
 __all__ = (
     *failures.__all__,
     
     *result.__all__,
-    *result_group.__all__,
 )