+++ vampytest/core/result/failures/asserting.py
@@ -3,7 +3,7 @@ __all__ = ('FailureAsserting',)
 from scarletio import copy_docs
 
 from .base import FailureBase
-from .helpers import add_documentation_into, add_route_parts_into
+from .helpers import render_documentation_into, render_route_parts_into, render_parameters_into
 
 
 class FailureAsserting(FailureBase):
@@ -48,12 +48,19 @@ class FailureAsserting(FailureBase):
     
     @copy_docs(FailureBase.get_failure_message)
     def get_failure_message(self):
+        handle = self.handle
         
         failure_message_parts = []
         failure_message_parts.append('Assertion failed at: ')
-        add_route_parts_into(self.handle, failure_message_parts)
+        failure_message_parts = render_route_parts_into(failure_message_parts, handle)
+        failure_message_parts = render_documentation_into(failure_message_parts, handle)
         
-        add_documentation_into(self.handle, failure_message_parts)
+        final_call_state = handle.final_call_state
+        if (final_call_state is not None) and final_call_state:
+            failure_message_parts.append('\nParameters: ')
+            failure_message_parts = render_parameters_into(
+                failure_message_parts, final_call_state.positional_parameters, final_call_state.keyword_parameters
+            )
         
         failure_message_parts.append('\n')
         self.assertion_exception.render_failure_message_parts_into(failure_message_parts)