+++ vampytest/core/result/failures/returning.py
@@ -3,7 +3,7 @@ __all__ = ('FailureReturning',)
 from scarletio import copy_docs
 
 from .base import FailureBase
-from .helpers import add_documentation_into, add_route_parts_into, render_parameters_into
+from .helpers import render_documentation_into, render_route_parts_into, render_parameters_into
 
 
 class FailureReturning(FailureBase):
@@ -14,9 +14,9 @@ class FailureReturning(FailureBase):
     ----------
     handle : ``Handle``
         The test's handle running the test.
-    expected_value : `None`, `Any`
+    expected_value : `None`, `object`
         The expected returned value.
-    received_value : `None`, `Any`
+    received_value : `None`, `object`
         The received returned value.
     """
     __slots__ = ('expected_value', 'received_value',)
@@ -29,9 +29,9 @@ class FailureReturning(FailureBase):
         ----------
         handle : ``Handle``
             The test's handle running the test.
-        expected_value : `None`, `Any`
+        expected_value : `None`, `object`
             The expected returned value.
-        received_value : `None`, `Any`
+        received_value : `None`, `object`
             The received returned value.
         """
         self = FailureBase.__new__(cls, handle)
@@ -56,16 +56,20 @@ class FailureReturning(FailureBase):
     
     @copy_docs(FailureBase.get_failure_message)
     def get_failure_message(self):
+        handle = self.handle
+        
         failure_message_parts = []
-        
         failure_message_parts.append('Unexpected return at: ')
-        add_route_parts_into(self.handle, failure_message_parts)
+        failure_message_parts = render_route_parts_into(failure_message_parts, handle)
+        failure_message_parts = render_documentation_into(failure_message_parts, handle)
         
-        add_documentation_into(self.handle, failure_message_parts)
         
-        
-        failure_message_parts.append('\nParameters: ')
-        render_parameters_into(self.handle.final_call_state, failure_message_parts)
+        final_call_state = handle.final_call_state
+        if (final_call_state is not None) and final_call_state:
+            failure_message_parts.append('\nParameters: ')
+            failure_message_parts = render_parameters_into(
+                failure_message_parts, final_call_state.positional_parameters, final_call_state.keyword_parameters
+            )
         
         failure_message_parts.append('\nExpected return: ')
         failure_message_parts.append(repr(self.expected_value))